Moved all Create logic to Create folder
